move scheduling to its own file
