+++ firmware/controllers/system/system.mk
@@ -5,5 +5,4 @@ SYSTEMSRC_CPP =	\
 	$(PROJECT_DIR)/controllers/system/efi_gpio.cpp \
 	$(PROJECT_DIR)/controllers/system/periodic_task.cpp \
 	$(PROJECT_DIR)/controllers/system/dc_motor.cpp \
-	
-	
+	$(PROJECT_DIR)/controllers/system/timer/scheduler.cpp \
